uploading 1.6.4...
git-svn-id: https://plugins.svn.wordpress.org/knews/trunk@976624 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ direct/knews_cron_do.php
@@ -217,7 +217,7 @@ if ($Knews_plugin) {
 
 		$result = $wpdb->query( $query );
 
-		unlink(KNEWS_DIR . '/tmp/lockfile.txt');
+		if (is_file(KNEWS_DIR . '/tmp/lockfile.txt')) unlink(KNEWS_DIR . '/tmp/lockfile.txt');
 		if ($fp) fclose($fp);
 		
 		$pend=true;
